Updated code to reflect changes to db in b6d23bc0e1730338dba51d60be6058d22a3a0754

+++ BL/ExpCatBL.cs
@@ -70,7 +70,7 @@ namespace BL
                                         DateTime.Today.ToString());
                 // Adds a new row to the table in the cache with the wanted Id and name
                 Cache.SDB.t_expenses_category.
-                    Addt_expenses_categoryRow((uint)this.ID, this.Name);
+                    Addt_expenses_categoryRow(this.ID, this.Name);
             }
         }
